Add a SQLite mode that enables the new native Android SQLite shadows
@SQLiteMode(LEGACY) will use the sqlite4java backed shadows.
@SQLiteMode(NATIVE) will use the new set of shadows backed by native Android code
from AOSP.

Also, add a shadow picker for the sqlite shadows that picks the shadows
based on the mode.

There is some logic added to RobolectricTestRunnner that treats the non-default
SQLite shadows (currently the native shadows) as custom shadows in order to perform
invokedynamic switchpoint invalidation. This allows switching between the shadows
within the same sandbox, avoiding the need to create a new Sandbox is the SQLite
mode is changed.

PiperOrigin-RevId: 391836377