Fix an ArrayIndexOutOfBoundsException in ByteBucketArray
The bucket id has four bits which has a maximum number 15(0xf). However, when
the leading bit is 1, Java will auto-cast the byte to a negative integer and
cause an ArrayIndexOutOfBoundsException. For example, previously, if the
'index' value is -34 (0xffffffde), the bucket index would be -3.

To solve this, convert the id to an unsigned integer can help Java treat it as
a positive number. Continuing the example above, with the unsigned byte
conversion to int, the bucket index is now 13.

PiperOrigin-RevId: 395403444